rename tests_config_manager prefix is used to find test php files, so move 'tests' qualifier to end of file with extension.
